requirements.txt
sycamore-ai
